requirements.txt
# Invoice Hub Backend Requirements
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0

# OCR & Google Drive
mindee==4.10.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.141.0
openpyxl==3.1.5

# Optional: For production deployment
# gunicorn==21.2.0
# redis==5.0.1
